+++ src/components/AddItem.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function AddItem() {
+  return (
+    <div>AddItem</div>
+  )
+}
+
+export default AddItem
