--- src/components/HeaderSection.css
@@ -1,22 +1,0 @@
-@media screen and (max-width: 768px){
-    #navbar{
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-        justify-content: flex-start;
-
-        /*to seprate navlinks from the navbar */
-        position:fixed;
-        top: 90px;
-        right:0px;
-        width:300px;
-        height:100vh;
-        background-color: rgb(245, 214, 128);
-        box-shadow: 0 40px 60px rgba(149, 145, 145, 0.1);
-        padding: 40px 0 0 10px;
-    }
-    #mobile{
-        display: block;
-    }
-    
-}
